Final Decision
Although it somewhat bothers that by letting the flow field continue, it ends up being one strand, it goes well with my abstract simulation of desert winds. It starts windy and uncontrollable, once pressed, the user has control of the wind, but once you click again, you release that control to the wild and the winds are everywhere.